refactor!: move camera and game_state to utils
