+++ alien_invasion/entities/sprites/player.py
@@ -7,8 +7,8 @@ import pygame
 from alien_invasion import settings
 from alien_invasion.entities.sprites import Sprite, SpriteAnimationFactory
 from alien_invasion.entities.sprites.bullet import BulletFactory
-from alien_invasion.game_state import GameState
-from alien_invasion.sprite_manager import SpritesManager
+from alien_invasion.utils.game_state import GameState
+from alien_invasion.utils.sprite_manager import SpritesManager
 from alien_invasion.utils.timer import Timer
 
 
@@ -59,7 +59,7 @@ class Player(Sprite):
 @dataclass
 class PlayerFactory:
     speed: int = 300
-    layer: int = settings.GameLayer.GROUND.value
+    layer: int = settings.Layer.ENTITIES.value
 
     def create(self, pos: pygame.Vector2) -> Player:
         return Player(
